gcs/builder: use the prev_hash instead of block hash as siphash key
In this commit, we modify the filter creation to use the prev-hash of a
block instead of the block hash. We do this as, using the block hash
would be incompatible with committing the filter since it isn't possible
to know the block hash before adding the filter, since it itself affects
the block hash. The prev_hash is known at the time of block construction
eliminating any circular dependencies.